py/requirements.txt
argparse>=1.4.0
py-color-log>=1.0.4
passlib>=1.7.4